extract useGameClock hook and buildDashboardProps selector
- hooks/useGameClock.js: client-side clock animation (rAF loop, pause
  sync refs, reset) moved out of App; App calls resetClock() on reset
- selectors/buildDashboardProps.js: the 225-line gameState->props
  mapping as a pure function; App keeps a thin useMemo call
- App.jsx: 972 -> 649 lines. Now: polling, routing, handlers, mounting

Closes #32
